requirements.txt
aiohttp==3.7.4.post0
aniso8601==9.0.1
async-timeout==3.0.1
attrs==21.2.0
blinker==1.4
boto3==1.17.97
botocore==1.20.97
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
EasyProcess==0.3
Flask==2.0.1
Flask-RESTful==0.3.9
grpcio==1.38.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
kubemq==2.2.4
MarkupSafe==2.0.1
multidict==5.1.0
pika==1.2.0
protobuf==3.17.3
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
PyVirtualDisplay==2.2
requests==2.25.1
s3transfer==0.4.2
selenium==3.141.0
sentry-sdk==1.3.0
six==1.16.0
slack-sdk==3.7.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
yarl==1.6.3
